Refactor: Decouple dns-store and dns serving
Decouple dns store, which is essentially a k-v store built upon Raft protocol,
  and dns serving logic, like handling non-recursive and recursive dns queries
  using dns store.
